+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## 2018-02-18 - 0.7.3
+## 2018-02-18 - 0.8.0
 
+- Implement `getStartCoord` & `getEndCoord` methods, easy way to extract coordinates from a GeoJSON LineString.
+- Implement `getFormOfWay` method, easy way to extract FormOfWay from a GeoJSON LineString.
 - Update `distanceToNextRef` to use `start` & `end` params.
 
 ## 2018-02-17